:recycle: Refactors how we are applying the API Key
Realized there was an issue with how we were applying the api_key to requests so this hopefully fixes that whole thing.

A side bonus we did here was to move the changelog out of the readme and into it's own thing. We'll cleanup the readme a bit better in a future release

On branch refactor-api-key-usage
 Changes to be committed:
	new file:   CHANGELOG.md
	modified:   README.md
	modified:   src/Client.php
	modified:   src/Resources/Postcards.php
	modified:   tests/ClientTest.php
	modified:   tests/Resources/AddressesTest.php
	modified:   tests/Resources/CampaignsTest.php
	modified:   tests/Resources/EventsTest.php
	modified:   tests/Resources/FilesTest.php
	modified:   tests/Resources/PostcardsTest.php
	modified:   tests/Resources/RecipientsTest.php
	modified:   tests/Resources/ReportingTest.php
	modified:   tests/Resources/SelectionsTest.php
	modified:   tests/Resources/ToolsTest.php